+++ migrations/incremental-20200519-region-workgroup-function.sql
@@ -1,0 +1,8 @@
+CREATE TABLE `fs_region_function`
+(
+ `id` INT(10) NOT NULL AUTO_INCREMENT ,
+ `region_id` INT NOT NULL,
+ `function_id` INT,
+  `target_id` INT,
+   PRIMARY KEY (`id`)
+) ;