+++ src/Dev/SeedCommand.php
@@ -8,6 +8,7 @@ use Codeception\Lib\Di;
 use Codeception\Lib\ModuleContainer;
 use Foodsharing\Modules\Core\DBConstants\Region\RegionIDs;
 use Foodsharing\Modules\Core\DBConstants\Region\Type;
+use Foodsharing\Modules\Core\DBConstants\Region\WorkgroupFunction;
 use Foodsharing\Modules\WorkGroup\WorkGroupGateway;
 use Helper\Foodsharing;
 use Symfony\Component\Console\Command\Command;
@@ -153,12 +154,31 @@ class SeedCommand extends Command implements CustomCommandInterface
 		$I->addRegionMember($ag_testimonials, $user2['id']);
 
 		// Make ambassador responsible for all work groups in the region
-		$this->output->writeln('- make ambassador responsible for all work groups');
+		$this->output->writeln('- make ambassador responsible for all work groups except welcome group');
+		$welcomeGroupIdRead = $I->grabColumnFromDatabase('fs_region_function', 'region_id', ['target_id' => $region1, 'function_id' => 1]);
 		$workGroupsIds = $I->grabColumnFromDatabase('fs_bezirk', 'id', ['parent_id' => $region1, 'type' => Type::WORKING_GROUP]);
 		foreach ($workGroupsIds as $id) {
-			$I->addRegionMember($id, $userbot['id']);
-			$I->addRegionAdmin($id, $userbot['id']);
+			if ($welcomeGroupIdRead) {
+				if ($welcomeGroupIdRead[0] == $id) {
+					continue;
+				}
+			} else {
+				$I->addRegionMember($id, $userbot['id']);
+				$I->addRegionAdmin($id, $userbot['id']);
+			}
 		}
+
+		// Create a welcome Group if it doesn't exist.
+		if (!$welcomeGroupIdRead) {
+			$this->output->writeln('- create welcome group');
+
+			$welcomeGroup = $I->createWorkingGroup('Begrüßung', ['parent_id' => $region1, 'email_name' => 'Begrüßung Göttingen', 'teaser' => 'Hier sind die Begrüßer für unseren Bezirk']);
+			$I->haveInDatabase('fs_region_function', ['region_id' => $welcomeGroup['id'], 'function_id' => WorkgroupFunction::WELCOME, 'target_id' => $region1]);
+			$welcomeGroupIdRead[0] = $welcomeGroup['id'];
+		}
+		$this->output->writeln('- make foodsaver responsible for welcome group');
+		$I->addRegionMember($welcomeGroupIdRead[0], $user2['id']);
+		$I->addRegionAdmin($welcomeGroupIdRead[0], $user2['id']);
 
 		// Create store team conversations
 		$this->output->writeln('- create store team conversations');