+++ tests/acceptance/FoodsaverCanJoinRegionCept.php
@@ -1,5 +1,7 @@
 <?php
 
+use Foodsharing\Modules\Core\DBConstants\Region\WorkgroupFunction;
+
 $I = new AcceptanceTester($scenario);
 
 $I->wantTo('Join another region');
@@ -9,6 +11,10 @@ $user = $I->createFoodsaver(null, ['verified' => 0]);
 $region = $I->createRegion(null, 0, 3);
 $ambassador = $I->createAmbassador(null, ['bezirk_id' => $region['id']]);
 $I->addRegionAdmin($region['id'], $ambassador['id']);
+$welcomeAdmin = $I->createFoodsaver(null, ['verified' => 0]);
+$welcomeGroup = $I->createWorkingGroup('Begrüßung', ['parent_id' => $region['id']]);
+$I->haveInDatabase('fs_region_function', ['region_id' => $welcomeGroup['id'], 'function_id' => WorkgroupFunction::WELCOME, 'target_id' => $region['id']]);
+$I->addRegionAdmin($welcomeGroup['id'], $welcomeAdmin['id']);
 
 $I->login($user['email']);
 /*
@@ -26,4 +32,5 @@ $I->click('#becomebezirkchooser-button');
 $I->waitForElementVisible('//a[contains(text(), "Neues Thema")]');
 
 $I->seeInDatabase('fs_foodsaver_has_bell', ['foodsaver_id' => $ambassador['id']]);
+$I->seeInDatabase('fs_foodsaver_has_bell', ['foodsaver_id' => $welcomeAdmin['id']]);
 $I->seeInDatabase('fs_foodsaver_has_bezirk', ['foodsaver_id' => $user['id'], 'bezirk_id' => $region['id']]);